Simplify Swift object init a bit
Don't do the eager init for DynamicObject as we can easily do it on access
(since the only access is through the subscript function under our control),
and remove the logic for optional numbers which was never actually used for
anything.

Initialize generic properties for other objects entirely from obj-c to
eliminate a bunch of extra wrappers and steps in the process.